ADL: Fix playback issues (#6)
Calling rewind() before update() causes access violation.

The driver channels are now cleared in rewind (callback 8).  Before this
change, values for previous subsongs could be held over into the new
subsong when the previous subsong had a channel active that the new
subsong did not leaving a dataptr value on an inactive channel which
will then be processed in executePrograms.

Both rewind and update perform a check to be sure no subsong < 2 can be
selected.

Removed the update call from the end of rewind.  This would cause update
to occur twice in the first update call.  This affects the test results
because there is now an extra r72.00 near the beginning since there are
now 2 updates where there was originally 1.

Also added loadcursubsong to act as a flag that indicates that rewind has
been called and cursubsong needs loaded the next time update is called.

Calling rewind(-1) is a special case that initializes the opl and sets
itself up to load subsong 2.
Note: EOBSOUND.ADL has a subsong 1.

update() checks the loadcursubsong flag and, when sent, calls
playSoundEffect for cursubsong (this used to be in rewind, but it caused
subsong 2 to load up while initializing. Now doing something like
rewind();
songlength(4);
will report the correct song length and not be messed up due to subsong
2 being prepared when it's not wanted.

This still allows multiple subsongs to play simultaneously (by channel
priority) as well as each subsong to be cycled consecutively for Winamp
(though I imagine songlength and seek would be/has been whacked in
Winamp).

Constructor doesn't need init call, it's done in rewind(-1) now.