Avoid undefined shift in hybrid.cpp
Left shift of negative values has undefined behavior
in C and C++. Replace it with a properly defined
equivalent expression.

No functional change. But note that the calculation
looks bogus and may be wrong.

Since I don't know what it's supposed to do I leave it
like that. Needs further investigation, and should be
fixed in a follow-up patch if necessary.