Remove static variables and simplify code in CadlibDriver::ChangePitch()
The method CadlibDriver::ChangePitch() in src/adlib.cpp
uses static variables to cache computed values, including
a pointer. This is a recipe for disaster: When the method
is called for different CadlibDriver objects, it can reuse
the cached value/pointer from the other object.

If such a cache is needed it must be stored in the object
instead. But here it doesn't make much sense to cache the
values at all. Simply drop the "optimization" to fix the
issue.

While at it, simplify the computation to make it more
readable and efficient (no change of the resulting values).