Fix some memory leaks in destructors
The destructors of CjbmPlayer fails to deallocate the
"sequences" array. The classes CmdiPlayer, CmusPlayer,
and CsopPlayer manually call the destructor of their
driver classes instead of freeing the driver. (The bug
in CadlibDriver which made it dangerous to free it has
been fixed in the previous commit.)

Add the missing delete statements.

(This doesn't address any memory leaks observable when
load() is called twice. I don't know whether that is
even allowed.)