Fix OOB reads in Cu6mPlayer::get_next_codeword() (issue #92)
During LZW decompression in src/u6m.cpp, the size of the input
buffer is not checked when reading input codewords. So an
out-of-bounds read access can be triggered simply by reading
a truncated Ultima 6 Music file.

Change the 2nd parameter of get_next_codeword() from "const
char *" to "data_block&" to make the buffer size available
to the method and add a proper size check. Also return failure
from lzw_decompress() when an input error is detected.

Fixes: https://github.com/adplug/adplug/issues/92