Fix null pointer dereferences (DoS) in Cd00Player::load() (issue #101)
When a .d00 file has a version 2 header, but its version field
indicates version 0 or 1, the code in src/d00.cppattempts to
access the version 0/1 header, but header1 is not set.
Dereferencing this uninitialized pointer may try to access an
arbitrary address if the object reuses a previously allocated
memory block, and if the block is zeroed it results in a null
pointer dereference.

Fix it by checking that the version number is between 2 and 4.

Bug: https://github.com/adplug/adplug/issues/101
Reported-by: Lionel Debroux <lionel_debroux@yahoo.fr>