Avoid OOB accesses when playeing .d00 files
In src/d00.cpp, offsets read from file data are used in many
places without checking to access the filedata array (often
via other pointers). This commit adds the missing checks.

In order to be accessible from other methods than load(),
filesize is promoted to a class member. A macro INDEX_OK()
is defined to hide the error-prone details like necessary
casts and pointer arithmetic from the methods and keep the
checks readable.

Checks are added in update(), rewind(), setvolume(), setfreq(),
and setinst(). Also improve handling of invalid subsong numbers
in rewind() while at it.

Note: The class uses a lot of misaligned pointers. This is not
portable even though they are cast back to unsigned char*
before dereferencing. The code should be rewritten without type
punning, but that is outside the scope of this commit.