Fix description and instrument names handling in CdfmLoader
In src/dfm.cpp, add checks for the length bytes of songinfo and
instrument names to load().

In src/dfm.h, use the right constructor for char arrays in
getdesc() and getinstrument(), and check for a valid instrument
number in the latter.

That prevents OOB reads or throwing std::out_of_range errors
when requesting these strings.