a2m.{h,cpp}: Move sixdepak functionality to a subclass
Move the class members and methods used for decoding compressed
files to a new subclass. This subclass is only instantiated
in static method sixdepak::decode() (formerly sixdepak()), so
decompression data doesn't clutter the player class. Note that
the sixdepak subclass could easily moved out of Ca2mLoader class
and made local to the .cpp file or moved to its own header.

Remove all preprocessor constants and const members along the
way and replace them with enum constants. Replace the lookup
tables with static member functions (one of them is still
implemented with a lookup table, but the other two are simpler
and faster with arithmetic).

The old decode() method is renamed to do_decode() and returns
the decoded length now. The name is now used for the entry point,
and the code from the old sixdepak() partially went into the
constructor, do_decode(), and the new decode(). It allocates
a sixdepak instance dynamically, so buf became an array an is no
longer allocated separately.

Note that output_size is now unused. Future changes should use
it for the size of the destination buffer instead of hardcoded
MAXBUF.

No functional changes.