cff.cpp: Check that number of patterns is valid
There is only enough space for 36 patterns. Trying to load more
will access memory past the end of the allocated arrays. Fix
that by adding a check that the number is valid (also rejecting
modules that are too short for the number of patterns or that
have no patterns).

This commit fixes OOB memory accesses while loading track data
for the following fuzzing samples:
  i-110_012.cff, i-110_013.cff, i-110_014.cff, i-110_015.cff,
  i-110_017.cff, i-110_018.cff, i-110_019.cff, i-110_020.cff,
  i-110_021.cff, i-110_022.cff, i-110_023.cff, i-110_024.cff,
  i-110_092.cff, i-110_093.cff, i-110_101.cff, i-110_102.cff,
  i-110_103.cff.

It also fixes the crashes in CmodPlayer::update() during playback
for i-110_016.cff and i-110_097.cff.

Bug: https://github.com/adplug/adplug/issues/110