amd.cpp: Check order length and number of patterns in load()
Both nop and length are set to values read from the file.
Check that they do not exceed the corresponding array sizes.

This fixes crashes while loading fuzzing samples i-110_033.amd
and i-110_072.amd (the latter is now loaded as a General MIDI
file and suffers from an integer overflow and endless playback).

Bug: https://github.com/adplug/adplug/issues/110