amd.cpp: Add two more missing checks
The order list is read from the file without checking. An
invalid pattern number can trigger OOB reads during playback.
Add a check that rejects the file if a number that could lead
to an OOB access is found. Note that while the pattern numbers
really should be < nop, we are a bit more forgiving and only
check for them being < 64 since that's the allocated length of
the arrays.

A few fuzzung samples no longer load or are recognized as
LucasArts AdLib MIDI now. That's fine, except that the midi
player has its own issues.

In the track loading section for unpacked modules, limit the
outer loop to t < maxi to avoid accessing the tracks array
past its end.