psi.{h,cpp}: Use separate position pointers
The pointers to the channel's positions are updated directly
in the tune data in xadplayer_update(). This is a questionable
approach. Moreover, they are not reset to the original values
in xadplayer_rewind(), so rewinding doesn't work as expected.

Use a separate copy of channel pointers which is initialized
in xadplayer_rewind() and modified in xadplayer_update() to
rectify the situation.