sa2.cpp: Check order list for invalid entries
If the order list contains an entry >= 64 it can lead to an OOB
access of the tracks array during playback. Add a check in load()
to prevent that (being more strict, enforce entries to be < nop).

Note that a similar check for track orders isn't necessary since
the number of allocated tracks is greater than the maximal value
of a byte.