mid.cpp: Fix an OOB read and very long runtime
Some of the fuzzing samples can be loaded by CmidPlayer, too.
So fix the problems that occur in that case as well.

Prevent accessing a lookup table with negative index and
improve excessively long running tests on broken files that
could require billions of update() calls to wait for the
next event:

* Ignore negative note values in midi_fm_playnote().
* Limit variable length numbers to allowed range.
* Fix wait calculation, initital value wasn't big enough.
* Limit track length to file size.