Re-enable TwinTeam test
The test passes consistently in my setup, and there doesn't seem to
be anything that's supposed to produce varying results in the player.
The inconsistent results reported earlier may have been caused by code
exhibiting undefined behaviour that has been fixed in the meantime
(perhaps commit 39d0799693b8, "Fix undefined behaviour"). If the test
still fails in some setups that's a bug that should be fixed instead
of hiding it.

This commit re-enables the TwinTrack sample in playertest, reverting
commit 95310fdb7271 ("Disable TwinTrack tests"), and removes it from
stresstest, where it was initially added as an example.