imf: Rewrite size calculations more clearly and cleanup in load()
Restructure code and add/rename variables to make the size
calculations more uniform, simpler, and easier to follow.

The validity checks have been changed to allow files that
end before the time value of the last sound command. Such
files seem to be somewhat common. Strictness of the checks
is now the same for either flavour of the format.