s3m, dmo: Move code to load a pattern in a new method, respect length
Pattern data has the same format in s3m (after the length word)
and in dmo (in the unpacked stream). Move the code to load a
pattern to a new method Cs3mPlayer::load_pattern() so it can be
reused for dmo.

Update the code to respect the length of the pattern data which
was ignored before.

Minimize code repetition by using a closure object (poor man's
lambda function for compatibility with old C++ standards) for
file reading+accounting, and a reference to the current event.