adl: Add crude loop detection
The adl driver doesn't detect looping songs and plays them
forever.

In theory, the file format allows for quite complex channel
control flow, and a reliable detection whether a song is
endlessly repeating previously played parts is not feasible.
In practice, for the music actually released in this format,
a simple check seems to be good enough (although it's easy to
craft malicious examples where this check fails; both false
positives and false negatives):

Let's assume a song is repeating if all channels which are
still playing have jumped back or are (transitively) waiting
on channels which have jumped back.

Bug: https://github.com/adplug/adplug/issues/122