adl: Fix rewind()
CadlPlayer::rewind() doesn't really do what it's supposed to.
For example, rewind(-1) starts subsong 2 instead of the current
one, with other arguments it overlays the new song without
stopping the old sounds, and an invalid subsong number will
cause a crash during update().

Fix these issues and also enqueue the subsong directly in
rewind() (there is no benefit in delaying that to update(), it
only made the code more complicated).