CI: Replace old Travis CI with GitHub Actions
We didn't have Travis CI working for some years,
which inevitably led to some problems on different platforms.
I have fixed some of them in previous commits, but not all.

Also I have tried to set up Travis again, but found it burdensome.
So I'm replacing it with a known built-in solution - GitHub Actions.