stresstest: Deal with SIP dropping library search path on macOS
On macOS, stresstest fails to re-execute itself for the individual
test cases when adplug is not already installed (and uses the wrong
library otherwise). It tries to load the library from the install
destination instead of the build directory. The reason is that
Apple's System Integrity Protection drops some environment variables,
including LD_LIBRARY_PATH and DYLD_LIBRARY_PATH, when it executes
system programs like the shell, which is used by system(). The
autotools build system relies on these variables to override the
library search path when it runs tests before installation.

As a workaround, restore the value of these environment variables
from within the shell by adding the variable assignment to the
constructed command line.

Revert previous workaround commit 3a90fc3d.

Closes: https://github.com/adplug/adplug/issues/138