stresstest: Limit sh style quoting to unix, add basic windows version
Quoting commands according to sh syntax may break command execution
on non-unix operating systems, so only do it on unix.

On Windows, there is no universal quoting rule, so add a simple
implementation that should be good enough to pass down arguments
with spaces and other special characters that may appear in file
names (but not newlines and maybe other strange chars). This will,
however, not work correctly in other contexts.

On other operating systems, no quoting is performed and a warning
is shown.