Fix RAD v2 nested riff precedence.
In RAD v2, when the first line of a riff references another riff,
the innermost riff is the riff that continues playing afterward.
The upstream RAD v2 replayer has a bug where the innermost riff
will be played, but then the outer riff(s) will overwrite the riff
track pointer as the recursive call unrolls. The discrepancy between
how the tracker and the replayer perform this seems to be due
to the tracker playing fully unpacked pattern data (thus it
does not need to store an updated track pointer on return).

I've fixed this bug by adding an updated flag to the CRiff struct.
When recursively invoking a riff, `TickRiff` will set this flag
to `false`, and when a riff is handled, `TickRiff` will set this
flag to `true`. If this flag is `true` when `TickRiff` is exiting
from a recursive call, it will skip the tasks at the end of
`TickRiff` (which are no longer relevant to this clobbered riff).
This is the least invasive solution I was able to come up with,
and while I can't guarantee it's 100% correct, it has fixed every
case I've tried so far.

Adapted from my patch to the MegaZeux fork of the RAD replayer
(currently hoping for feedback from Shayde):
https://github.com/AliceLR/megazeux/pull/488